- Re added stacktrace copy when click isnt double in error screen - Fixed various Block events - Fixed InventoryAttachment related classes - Fixed GameRulesMixin & GameRulesKJS preserving Key - Fixed FluidIngredientKJS and added SizedFluidIngredient override - Fixed 15 other classes

+++ src/main/java/dev/latvian/mods/kubejs/block/FarmlandTrampledKubeEvent.java
@@ -20,7 +20,7 @@ public class FarmlandTrampledKubeEvent implements KubeEntityEvent {
 	}
 
 	@Info("The distance of the entity from the block.")
-	public float getDistance() {
+	public double getDistance() {
 		return event.getFallDistance();
 	}
